LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/mathml/relations/css-styling/floats/not-floating-001-expected.txt


FAIL float child ignored in maction assert_approx_equals: inline position (child 1) expected -28 +/- 1 but got -8
PASS float child ignored in menclose
PASS float child ignored in merror
PASS float child ignored in mfrac
FAIL float child ignored in mi assert_approx_equals: inline size expected 82.90625 +/- 1 but got 55.8125
PASS float child ignored in mmultiscripts
FAIL float child ignored in mn assert_approx_equals: inline size expected 81.3125 +/- 1 but got 54.21875
FAIL float child ignored in mo assert_approx_equals: inline size expected 90.1875 +/- 1 but got 63.09375
PASS float child ignored in mover
PASS float child ignored in mpadded
PASS float child ignored in mphantom
PASS float child ignored in mroot
PASS float child ignored in mrow
FAIL float child ignored in ms assert_approx_equals: inline size expected 81.3125 +/- 1 but got 54.21875
PASS float child ignored in msqrt
PASS float child ignored in mstyle
PASS float child ignored in msub
PASS float child ignored in msubsup
PASS float child ignored in msup
FAIL float child ignored in mtext assert_approx_equals: inline size expected 81.3125 +/- 1 but got 54.21875
PASS float child ignored in munder
PASS float child ignored in munderover
FAIL float child ignored in semantics assert_approx_equals: inline position (child 1) expected -28 +/- 1 but got -8
